requirements.txt
flask
pika
xmltodict
python-dateutil